Move param logic to test file
